+++ includes/admin-metaboxes.php
@@ -694,16 +694,8 @@
         <ul>
             <li><a href="https://wordpress.org/support/view/plugin-reviews/wp-rss-aggregator?rate=5#postform" target="_blank"><?php _e( 'Give it a 5 star rating on WordPress.org', WPRSS_TEXT_DOMAIN ) ?></a></li>
         </ul>
-        <p><strong><?php _e( 'Add functionality with our premium extensions:', WPRSS_TEXT_DOMAIN ) ?></strong></p>
-        <?php $addons = wprss_addons_get_extra() ?>
-        <?php if (count($addons)): ?>
-        <ul class="add-on-list">
-            <?php foreach ($addons as $_code => $_addon): ?>
-            <li class="add-on <?php echo sprintf('add-on-code-%1$s', $_code) ?>"><a href="<?php echo $_addon['url'] ?>"><?php echo $_addon['title'] ?></a></li>
-            <?php endforeach ?>
-        </ul>
-        <?php endif ?>
         <?php
+        do_action('wpra_share_the_love_metabox');
     }