+++ src/Logger/WpdbLogger.php
@@ -55,11 +55,11 @@ class WpdbLogger extends AbstractLogger implements ClearableLoggerInterface, Log
     /**
      * A mapping of log properties to table column names.
      *
-     * @see   COL_DATE
-     * @see   COL_LEVEL
-     * @see   COL_MESSAGE
-     *
      * @since 4.13
+     *
+     * @see   WpdbLogger::LOG_DATE
+     * @see   WpdbLogger::LOG_LEVEL
+     * @see   WpdbLogger::LOG_MESSAGE
      *
      * @var string[]
      */
@@ -83,10 +83,12 @@ class WpdbLogger extends AbstractLogger implements ClearableLoggerInterface, Log
      * @param array          $columns A mapping of log data keys to column names. The `COL_*` constants may be used to
      *                                map standard log properties and any non-standard properties may also be included
      *                                to insert fixed values with the $extra parameter.
-     * @param array          $extra   A mapping of log properties and the values to insert for them. Beware that the
-     *                                standard log properties MAY be overridden if specified as keys for this
-     *                                parameter. All non-standard data is stored in the table as VARCHAR with a limit
-     *                                of 100 characters.
+     * @param array          $extra   A mapping of log properties and the values to insert for them. The values may be
+     *                                functions, which will be invoked during insertion. The function will receive the
+     *                                level, message and this $extra argument array as arguments.
+     *                                Be aware that the standard log properties, "id", "level", "message" and "date",
+     *                                will be overridden if those keys are given in this array.All non-standard data
+     *                                is stored in the table as VARCHAR with a limit of 100 characters.
      */
     public function __construct(TableInterface $table, $columns = [], $extra = [])
     {
@@ -203,16 +205,22 @@ class WpdbLogger extends AbstractLogger implements ClearableLoggerInterface, Log
         }
 
         if ($prop === static::LOG_LEVEL) {
-            return $level;
+            return (string) $level;
         }
 
         if ($prop === static::LOG_MESSAGE) {
             return $message;
         }
 
-        return isset($this->extra[$prop])
-            ? $this->extra[$prop]
-            : null;
+        if (!isset($this->extra[$prop])) {
+            return null;
+        }
+
+        if (is_callable($this->extra[$prop])) {
+            return call_user_func_array($this->extra[$prop], [$level, $message, $this->extra]);
+        }
+
+        return $this->extra[$prop];
     }
 
     /**